refactor(chansey): migrate profile page to settings with modular sub-components
- Move profile components (change-password, exchange-integrations, exchange-key-form, profile-info) from profile/ to settings/
- Add new settings sub-components: account, appearance, notifications, security, trading
- Extract settings constants and types into dedicated files
- Simplify settings parent component by delegating to child components
- Remove deprecated profile page, service, and types
- Update routes and topbar navigation to use settings path
- Add auto-save utility with debounced save and status tracking
- Add SaveStatusIndicatorComponent for inline save feedback replacing toast messages
- Persist panel collapsed state in localStorage via panel-state utility
- Integrate auto-save and panel persistence into notification and trading settings
- Extract createPanelState and createSinglePanelState factory functions to reduce boilerplate
- Add ChangeDetectionStrategy.OnPush to change-password, exchange-integrations, exchange-key-form, and profile-info
- Remove redundant standalone: true (Angular 20 default)
- Replace unsafe Record<string, unknown> casts with typed access via IUser.coinRisk and IUserProfileUpdate
- Replace setTimeout with RxJS timer + takeUntilDestroyed for proper cleanup in account-settings
- Add lazy loading to settings tabs for deferred rendering
- Add aria-label attributes and label-for associations for accessibility
- Fix standalone ngModel warning in security-settings disable-2FA dialog
