--- apps/chansey/src/app/pages/user/profile/index.ts
@@ -1,1 +1,0 @@
-export * from './profile.component';
