--- apps/chansey/src/app/pages/user/profile/profile.component.html
@@ -1,37 +1,0 @@
-@if (isLoading()) {
-  <app-profile-info [messages]="messages()" />
-} @else if (user(); as userData) {
-  <app-profile-info
-    #profileInfo
-    [user]="userData"
-    [risks]="risksQuery.data()"
-    [risksLoading]="risksQuery.isLoading() || risksQuery.isPending()"
-    [risksError]="risksQuery.isError()"
-    [isUpdating]="updateProfileMutation.isPending()"
-    [isUploadingImage]="uploadProfileImageMutation.isPending()"
-    [messages]="messages()"
-    (submitProfile)="onSubmitProfile($event)"
-    (uploadImage)="onUploadImage($event)"
-  />
-
-  <app-change-password
-    #changePasswordRef
-    [isPending]="changePasswordMutation.isPending()"
-    (changePassword)="onChangePassword($event)"
-  />
-
-  <app-exchange-integrations
-    [exchanges]="supportedExchangesQuery.data() || []"
-    [exchangeForms]="exchangeForms()"
-    [isLoading]="supportedExchangesQuery.isLoading() || supportedExchangesQuery.isPending()"
-    [isError]="supportedExchangesQuery.isError()"
-    [user]="userData"
-    (saveKeys)="onSaveExchangeKeys($event)"
-    (editKeys)="onEditExchangeKeys($event)"
-    (cancelEditKeys)="onCancelEditExchangeKeys($event)"
-    (removeKeys)="onRemoveExchangeKeys($event)"
-  />
-}
-
-<p-toast />
-<p-confirmDialog [style]="{ width: '450px' }" />
