+++ apps/chansey/src/app/pages/user/settings/components/change-password/change-password.component.html
@@ -1,12 +1,9 @@
-<p-card class="mx-auto mt-8 max-w-4xl p-4">
-  <ng-template pTemplate="header">
-    <div class="flex items-center gap-3 border-b border-gray-100 p-4 dark:border-neutral-700">
-      <i class="pi pi-lock text-primary text-2xl!"></i>
-      <div>
-        <h2 class="m-0 text-xl font-semibold">Change Password</h2>
-        <p class="text-surface-500 m-0 mt-1 text-sm">Update your account password</p>
-      </div>
-    </div>
+<p-panel [toggleable]="true" [collapsed]="panelCollapsed" (onAfterToggle)="onPanelToggle($event)">
+  <ng-template #header>
+    <span class="flex items-center gap-2">
+      <i class="pi pi-lock"></i>
+      <span class="font-semibold">Change Password</span>
+    </span>
   </ng-template>
   <p-fluid>
     <form [formGroup]="passwordForm" (ngSubmit)="onChangePassword()" class="mx-auto flex flex-col gap-4">
@@ -88,4 +85,4 @@
       </div>
     </form>
   </p-fluid>
-</p-card>
+</p-panel>
