+++ apps/chansey/src/app/pages/user/settings/components/notification-settings/notification-settings.component.html
@@ -1,0 +1,156 @@
+@if (notifPrefsQuery.isPending()) {
+  <!-- Notification Channels skeleton -->
+  <div class="border-surface-200 dark:border-surface-700 rounded-lg border p-4">
+    <p-skeleton width="180px" height="1.5rem" styleClass="mb-4" />
+    @for (i of [1, 2, 3]; track i) {
+      <div class="flex items-center justify-between py-3">
+        <div>
+          <p-skeleton width="160px" height="1.25rem" styleClass="mb-1" />
+          <p-skeleton width="240px" height="1rem" />
+        </div>
+        <p-skeleton width="3rem" height="1.5rem" borderRadius="1rem" />
+      </div>
+    }
+  </div>
+  <!-- Event Preferences skeleton -->
+  <div class="border-surface-200 dark:border-surface-700 mt-3 rounded-lg border p-4">
+    <p-skeleton width="160px" height="1.5rem" styleClass="mb-4" />
+    @for (i of [1, 2, 3, 4]; track i) {
+      <div class="flex items-center justify-between py-2">
+        <div>
+          <p-skeleton width="140px" height="1.1rem" styleClass="mb-1" />
+          <p-skeleton width="200px" height="0.85rem" />
+        </div>
+        <p-skeleton width="3rem" height="1.5rem" borderRadius="1rem" />
+      </div>
+    }
+  </div>
+  <!-- Quiet Hours skeleton -->
+  <div class="border-surface-200 dark:border-surface-700 mt-3 rounded-lg border p-4">
+    <p-skeleton width="120px" height="1.5rem" styleClass="mb-4" />
+    <div class="flex items-center justify-between py-3">
+      <div>
+        <p-skeleton width="160px" height="1.25rem" styleClass="mb-1" />
+        <p-skeleton width="280px" height="1rem" />
+      </div>
+      <p-skeleton width="3rem" height="1.5rem" borderRadius="1rem" />
+    </div>
+  </div>
+} @else {
+  <div class="mb-3 flex justify-end">
+    <app-save-status [status]="autoSave.status()" />
+  </div>
+  <p-panel
+    [toggleable]="true"
+    [collapsed]="panelCollapsed['channels']"
+    (onAfterToggle)="onPanelToggle('channels', $event)"
+  >
+    <ng-template #header>
+      <span class="flex items-center gap-2">
+        <i class="pi pi-envelope"></i>
+        <span class="font-semibold">Notification Channels</span>
+      </span>
+    </ng-template>
+    <div class="flex flex-col gap-4">
+      <div class="flex flex-row items-center justify-between py-3">
+        <div>
+          <div class="text-900 mb-1 text-lg font-semibold">Email Notifications</div>
+          <div class="text-600 text-sm">Receive trade alerts and summaries via email</div>
+        </div>
+        <p-toggleswitch [(ngModel)]="prefChannelEmail" (onChange)="saveNotificationPreferences()" />
+      </div>
+      <div class="flex flex-row items-center justify-between py-3">
+        <div>
+          <div class="text-900 mb-1 text-lg font-semibold">Push Notifications</div>
+          <div class="text-600 text-sm">Browser push notifications for real-time alerts</div>
+        </div>
+        <p-toggleswitch [(ngModel)]="prefChannelPush" (onChange)="togglePushNotifications($event)" />
+      </div>
+      <div class="flex flex-row items-center justify-between py-3">
+        <div>
+          <div class="text-900 mb-1 text-lg font-semibold">SMS Notifications</div>
+          <div class="text-600 text-surface-400 text-sm">Coming soon</div>
+        </div>
+        <p-toggleswitch [ngModel]="false" [disabled]="true" />
+      </div>
+    </div>
+  </p-panel>
+
+  <p-panel
+    [toggleable]="true"
+    [collapsed]="panelCollapsed['events']"
+    (onAfterToggle)="onPanelToggle('events', $event)"
+    class="mt-3"
+  >
+    <ng-template #header>
+      <span class="flex items-center gap-2">
+        <i class="pi pi-list"></i>
+        <span class="font-semibold">Event Preferences</span>
+      </span>
+    </ng-template>
+    <div class="flex flex-col gap-3">
+      @for (evt of notificationEventOptions; track evt.key) {
+        <div class="flex flex-row items-center justify-between py-2">
+          <div>
+            <div class="text-900 font-semibold">{{ evt.label }}</div>
+            <div class="text-600 text-xs">{{ evt.description }}</div>
+          </div>
+          <p-toggleswitch [(ngModel)]="evt.enabled" (onChange)="saveEventPreference()" />
+        </div>
+      }
+    </div>
+  </p-panel>
+
+  <p-panel
+    [toggleable]="true"
+    [collapsed]="panelCollapsed['quietHours']"
+    (onAfterToggle)="onPanelToggle('quietHours', $event)"
+    class="mt-3"
+  >
+    <ng-template #header>
+      <span class="flex items-center gap-2">
+        <i class="pi pi-moon"></i>
+        <span class="font-semibold">Quiet Hours</span>
+      </span>
+    </ng-template>
+    <div class="flex flex-col gap-4">
+      <div class="flex flex-row items-center justify-between py-3">
+        <div>
+          <div class="text-900 mb-1 text-lg font-semibold">Enable Quiet Hours</div>
+          <div class="text-600 text-sm">Suppress non-critical notifications during specified hours (UTC)</div>
+        </div>
+        <p-toggleswitch [(ngModel)]="prefQuietEnabled" (onChange)="saveNotificationPreferences()" />
+      </div>
+      @if (prefQuietEnabled) {
+        <div class="flex items-center gap-4">
+          <div class="flex items-center gap-2">
+            <label for="quietHoursFrom" class="text-900 font-semibold">From</label>
+            <select
+              id="quietHoursFrom"
+              class="border-surface rounded-lg border px-3 py-2"
+              [(ngModel)]="prefQuietStart"
+              (ngModelChange)="saveNotificationPreferences()"
+            >
+              @for (h of hours; track h) {
+                <option [ngValue]="h">{{ h }}:00 UTC</option>
+              }
+            </select>
+          </div>
+          <div class="flex items-center gap-2">
+            <label for="quietHoursTo" class="text-900 font-semibold">To</label>
+            <select
+              id="quietHoursTo"
+              class="border-surface rounded-lg border px-3 py-2"
+              [(ngModel)]="prefQuietEnd"
+              (ngModelChange)="saveNotificationPreferences()"
+            >
+              @for (h of hours; track h) {
+                <option [ngValue]="h">{{ h }}:00 UTC</option>
+              }
+            </select>
+          </div>
+        </div>
+      }
+    </div>
+  </p-panel>
+}
